Separation of the admin screen into components
