+++ src/components/administrator/scss/AdministratorPanel.scss
@@ -1,0 +1,263 @@
+main {
+    margin-top: 1.5rem;
+
+    overflow: hidden;
+
+    max-width: 100%;
+    height: calc(100vh - 24px);
+
+    color: var(--light-grey2);
+
+    overflow: hidden;
+
+    background-color: var(--grey-dark2);
+    //border: solid 5px pink;
+    border-top-left-radius: 10px;
+
+    transition: 0.2s ease-out;
+
+    .panel-navigation {
+        display: flex;
+        width: 100%;
+        height: 15%;
+        padding-bottom: 15px;
+
+        justify-content: center;
+        align-items: flex-end;
+
+        button {
+            font-weight: 500;
+
+            height: 40%;
+
+            width: 100%;
+
+            text-transform: uppercase;
+
+            color: var(--light-grey2);
+            background-color: var(--dark-alt);
+            font-weight: 400;
+            box-shadow: 0 5px 30px var(--shadow);
+
+            &:hover {
+                background-color: var(--dark-alt2);
+                border: solid 3px var(--dark-alt);
+
+                color: var(--primary);
+            }
+        }
+
+        .active {
+            background-color: var(--dark-alt2);
+            border: solid 3px var(--dark-alt);
+
+            color: var(--primary);
+        }
+    }
+
+    .btn-left-bordered {
+        border-top-left-radius: 15px;
+        border-end-start-radius: 15px;
+        border-right: 1px solid rgba(255, 255, 255, 0.2);
+    }
+
+    .btn-right-bordered {
+        border-end-end-radius: 15px;
+        border-start-end-radius: 15px;
+        border-left: 1px solid rgba(255, 255, 255, 0.2);
+    }
+
+    .div-container {
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-start;
+        align-items: center;
+
+        padding: 15px;
+
+        height: 85%;
+        max-width: 100%;
+
+        //border: solid 5px green;
+
+        overflow: hidden;
+
+        .div-panel {
+            width: 60%;
+            height: 100%;
+            background-color: var(--grey-mid);
+
+            border-radius: 15px;
+            box-shadow: 0 5px 30px var(--shadow);
+            display: flex;
+            flex-direction: column;
+
+            align-items: center;
+            justify-content: flex-start;
+            padding: 4rem;
+
+            .textfield {
+                width: 80%;
+                display: flex;
+                flex-direction: column;
+                align-content: center;
+                justify-content: center;
+
+                //border: solid 5px purple;
+
+                input {
+                    border: none;
+                    border-radius: 10px;
+                    padding: 10px;
+                    background: var(--dark-alt);
+                    color: var(--light-grey2);
+                    font-size: 12pt;
+                    box-shadow: 0 10px 30px var(--shadow);
+                    outline: none;
+                    box-sizing: border-box;
+
+                    box-shadow: 0 5px 30px var(--shadow);
+                }
+
+                input::placeholder {
+                    color: var(--light-grey);
+                }
+            }
+
+            .textfield-buttons {
+                display: flex;
+                width: 80%;
+
+                align-content: center;
+                justify-content: center;
+
+                button {
+                    width: 100%;
+
+                    text-transform: uppercase;
+
+                    font-weight: 500;
+
+                    box-shadow: 0 5px 30px var(--shadow);
+                }
+            }
+
+            .textfield-access {
+                margin: 2rem 0px;
+            }
+
+            .textfield-buttons-access {
+                margin: 2rem 0px;
+            }
+        }
+
+        .button-std {
+            font-weight: 500;
+
+            height: 3rem;
+
+            width: 100%;
+
+            text-transform: uppercase;
+
+            color: var(--light-grey2);
+            background-color: var(--dark-alt);
+            font-weight: 400;
+            box-shadow: 0 5px 30px var(--shadow);
+
+            &:hover {
+                background-color: var(--dark-alt2);
+                border: solid 3px var(--dark-alt);
+
+                color: var(--primary);
+            }
+        }
+
+        .active {
+            background-color: var(--dark-alt2);
+            border: solid 3px var(--dark-alt);
+
+            color: var(--primary);
+        }
+
+        .container-subpanel-invite {
+            display: flex;
+            flex-direction: column;
+
+            justify-content: center;
+            align-items: center;
+
+            box-shadow: 0 5px 30px var(--shadow);
+
+            margin: 1rem 0px;
+
+            border-radius: 15px;
+            overflow: hidden;
+
+            width: 80%;
+
+            background-color: var(--dark-alt);
+
+            overflow: auto;
+
+            .textfield-buttons-invites {
+                margin-bottom: 2rem 0rem;
+                width: 100%;
+            }
+
+            .panel-create-invite {
+                display: flex;
+                flex-wrap: wrap;
+                justify-content: center;
+                align-items: center;
+
+                margin-top: 0.5rem;
+
+                justify-content: center;
+                align-items: center;
+                text-align: center;
+
+                width: 100%;
+                height: 100%;
+
+                button {
+                    width: 100%;
+                }
+
+                overflow-y: auto;
+
+                .code-invite {
+                    border-bottom: 1px solid rgba(255, 255, 255, 0.2);
+                    //border: solid 5px pink;
+                    width: 100%;
+
+                    text-align: center;
+
+                    border-radius: 10px;
+
+                    margin-bottom: 2rem;
+                }
+
+                .div-invites-search {
+                    display: flex;
+                    width: 100%;
+                    padding-left: 1rem;
+                    padding-top: 0.5rem;
+                    margin-right: 0.5rem;
+
+                    border-radius: 10px;
+                    border-bottom: 1px solid rgba(255, 255, 255, 0.2);
+
+                    .flex {
+                        flex: 1 1 0;
+                    }
+
+                    button {
+                        width: 6rem;
+                        margin-bottom: 8px;
+                    }
+                }
+            }
+        }
+    }
+}
